+++ lib/active_record_where_assoc/core_logic.rb
@@ -13,6 +13,7 @@ module ActiveRecordWhereAssoc
     # => "EXISTS (SELECT... *relation1*) OR EXISTS (SELECT... *relation2*)"
     def self.sql_for_any_exists(relations)
       relations = [relations] unless relations.is_a?(Array)
+      relations = relations.reject { |rel| rel.is_a?(ActiveRecord::NullRelation) }
       sqls = relations.map { |rel| "EXISTS (#{rel.select('1').to_sql})" }
       if sqls.size > 1
         "(#{sqls.join(" OR ")})" # Parens needed when embedding the sql in a `where`, because the OR could make things wrong
@@ -32,6 +33,7 @@ module ActiveRecordWhereAssoc
     # => "SUM((SELECT... *relation1*)) + SUM((SELECT... *relation2*))"
     def self.sql_for_sum_of_counts(relations)
       relations = [relations] unless relations.is_a?(Array)
+      relations = relations.reject { |rel| rel.is_a?(ActiveRecord::NullRelation) }
       # Need the double parentheses
       relations.map { |rel| "SUM((#{rel.to_sql}))" }.join(" + ").presence || "0"
     end
@@ -83,7 +85,7 @@ module ActiveRecordWhereAssoc
       end
 
       nested_relations = relations_on_association(record_class, association_names, given_conditions, options, deepest_scope_mod, NestWithSumBlock)
-
+      nested_relations = nested_relations.reject { |rel| rel.is_a?(ActiveRecord::NullRelation) }
       nested_relations.map { |nr| "COALESCE((#{nr.to_sql}), 0)" }.join(" + ").presence || "0"
     end
 
@@ -322,6 +324,9 @@ module ActiveRecordWhereAssoc
       if user_defined_actual_source_reflection(reflection).macro == :belongs_to || option_value(options, :ignore_limit)
         return current_scope.unscope(:limit, :offset, :order)
       end
+
+      # No need to do transformations if this is already a NullRelation
+      return current_scope if current_scope.is_a?(ActiveRecord::NullRelation)
 
       current_scope = current_scope.limit(1) if reflection.macro == :has_one